Remove ambiguity in requestFrom(uint8_t, int) signature

+++ src/BBQ10Keyboard.cpp
@@ -170,7 +170,7 @@ uint8_t BBQ10Keyboard::readRegister8(uint8_t reg) const
     m_wire->write(reg);
     m_wire->endTransmission();
 
-    m_wire->requestFrom(m_addr, 1);
+    m_wire->requestFrom((int)m_addr, (int)1);
     if (m_wire->available() < 1)
         return 0;
 
@@ -183,7 +183,7 @@ uint16_t BBQ10Keyboard::readRegister16(uint8_t reg) const
     m_wire->write(reg);
     m_wire->endTransmission();
 
-    m_wire->requestFrom(m_addr, 2);
+    m_wire->requestFrom((int)m_addr, (int)2);
     if (m_wire->available() < 2)
         return 0;
